BLD Split src/js/esbuild.config.mjs into multiple files
This separates the build step for _pyodide.out.js and pyodide.js. They happen at
different steps so it makes more sense to keep them separate.
