+++ src/js/package.json
@@ -79,7 +79,8 @@
     "ws": false
   },
   "scripts": {
-    "build": "tsc --noEmit && node esbuild.config.mjs",
+    "build-inner": "node esbuild.config.inner.mjs",
+    "build": "tsc --noEmit && node esbuild.config.outer.mjs",
     "test": "npm-run-all test:*",
     "test:unit": "cross-env TEST_NODE=1 ts-mocha --node-option=experimental-loader=./test/loader.mjs --node-option=experimental-wasm-stack-switching -p tsconfig.test.json \"test/unit/**\"",
     "test:node": "cross-env TEST_NODE=1 mocha test/integration/**/*.test.js",
